+++ backend/app/data_output_configuration/snowflake/model.py
@@ -1,22 +1,20 @@
+from sqlalchemy import String
 from sqlalchemy.orm import Mapped, mapped_column
 
-from app.data_output_configuration.base_model import BaseDataOutputConfiguration
+from app.data_output_configuration.base_model import BaseTechnicalAssetConfiguration
 
 
-class SnowflakeDataOutput(BaseDataOutputConfiguration):
-    database: Mapped[str] = mapped_column(nullable=True, use_existing_column=True)
-    schema: Mapped[str] = mapped_column(nullable=True, use_existing_column=True)
-    table: Mapped[str] = mapped_column(nullable=True, use_existing_column=True)
-    bucket_identifier: Mapped[str] = mapped_column(
-        nullable=True, use_existing_column=True
-    )
-    database_path: Mapped[str] = mapped_column(nullable=True, use_existing_column=True)
-    table_path: Mapped[str] = mapped_column(nullable=True, use_existing_column=True)
-    access_granularity: Mapped[str] = mapped_column(
-        nullable=True,
-        use_existing_column=True,
-    )
+class SnowflakeTechnicalAssetConfiguration(BaseTechnicalAssetConfiguration):
+    __tablename__ = "snowflake_technical_asset_configurations"
+
+    database: Mapped[str] = mapped_column(String, nullable=True)
+    schema: Mapped[str] = mapped_column(String, nullable=True)
+    table: Mapped[str] = mapped_column(String, nullable=True)
+    bucket_identifier: Mapped[str] = mapped_column(String, nullable=True)
+    database_path: Mapped[str] = mapped_column(String, nullable=True)
+    table_path: Mapped[str] = mapped_column(String, nullable=True)
+    access_granularity: Mapped[str] = mapped_column(String, nullable=True)
+
     __mapper_args__ = {
-        "polymorphic_identity": "SnowflakeDataOutput",
-        "polymorphic_load": "inline",
+        "polymorphic_identity": "SnowflakeTechnicalAssetConfiguration",
     }