+++ backend/tests/factories/s3_data_output.py
@@ -1,11 +1,11 @@
 import factory
 
-from app.data_output_configuration.s3.model import S3DataOutput
+from app.data_output_configuration.s3.model import S3TechnicalAssetConfiguration
 
 
 class S3DataOutputFactory(factory.alchemy.SQLAlchemyModelFactory):
     class Meta:
-        model = S3DataOutput
+        model = S3TechnicalAssetConfiguration
 
     id = factory.Faker("uuid4")
     bucket = "bucket"